requirements.txt
llama-index==0.4.24
tiktoken==0.3.0
langchain==0.0.106
python-dotenv==1.0.0
flask